requirements.txt
Django==1.8
djangorestframework==3.6.4
